move AbstractTheme and AbstractCoreTheme to ExtensionsModule
