+++ docs/Themes/Class.md
@@ -4,7 +4,7 @@ Filename: `<Vendor><Name>Theme.php`
 
 Status: Required
 
-Must: `extends Zikula\ThemeModule\AbstractTheme`
+Must: `extends Zikula\ExtensionsModule\AbstractTheme`
 
 Description: This class establishes the extension in the Symfony ecosystem as a **Bundle**.
 See the parent classes for more information. This class does not require any content, though overriding certain
