+++ src/Zikula/CoreBundle/HttpKernel/ZikulaHttpKernelInterface.php
@@ -18,7 +18,7 @@ use Symfony\Component\HttpKernel\KernelInterface;
 use Symfony\Component\HttpKernel\RebootableInterface;
 use Symfony\Component\HttpKernel\TerminableInterface;
 use Zikula\ExtensionsModule\AbstractModule;
-use Zikula\ThemeModule\AbstractTheme;
+use Zikula\ExtensionsModule\AbstractTheme;
 
 interface ZikulaHttpKernelInterface extends KernelInterface, TerminableInterface, RebootableInterface
 {
