+++ src/Zikula/CoreBundle/HttpKernel/ZikulaKernel.php
@@ -25,6 +25,7 @@ use Zikula\BootstrapTheme\ZikulaBootstrapTheme;
 use Zikula\CategoriesModule\ZikulaCategoriesModule;
 use Zikula\ExtensionsModule\AbstractExtension;
 use Zikula\ExtensionsModule\AbstractModule;
+use Zikula\ExtensionsModule\AbstractTheme;
 use Zikula\ExtensionsModule\ZikulaExtensionsModule;
 use Zikula\GroupsModule\ZikulaGroupsModule;
 use Zikula\MailerModule\ZikulaMailerModule;
@@ -36,7 +37,6 @@ use Zikula\RssTheme\ZikulaRssTheme;
 use Zikula\SearchModule\ZikulaSearchModule;
 use Zikula\SecurityCenterModule\ZikulaSecurityCenterModule;
 use Zikula\SettingsModule\ZikulaSettingsModule;
-use Zikula\ThemeModule\AbstractTheme;
 use Zikula\ThemeModule\EventListener\AddJSConfigListener;
 use Zikula\ThemeModule\ZikulaThemeModule;
 use Zikula\UsersModule\ZikulaUsersModule;
