+++ src/system/AtomTheme/ZikulaAtomTheme.php
@@ -15,7 +15,7 @@ namespace Zikula\AtomTheme;
 
 use Symfony\Component\HttpFoundation\Response;
 use Twig\Environment;
-use Zikula\ThemeModule\AbstractCoreTheme;
+use Zikula\ExtensionsModule\AbstractCoreTheme;
 
 class ZikulaAtomTheme extends AbstractCoreTheme
 {
