+++ src/system/ExtensionsModule/AbstractCoreTheme.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Zikula\ThemeModule;
+namespace Zikula\ExtensionsModule;
 
 abstract class AbstractCoreTheme extends AbstractTheme
 {
